Move TableFreeze to new table/ directory
